Updated Docker integration tests.
Updated definitions.json to make the username and password consistent with the integration tests.

Updated images from microsoft/dotnet:2.2-sdk to mcr.microsoft.com/dotnet/sdk:8.0 and rabbitmq:3-management to rabbitmq:3.12.12-management.

Removed net472 as a Framework as it prevents projects from running in a Linux container. Given netstandard2.0 is supported, it's superfluous.